+++ GameCore/Render/RenderMaterial/MaterialManager.cs
@@ -39,7 +39,7 @@ namespace GameCore.Render.RenderMaterial
             return null;
         }
 
-        public ObjMaterial GetFromFile(ShaderProgram program, string aName, string aFileName)
+        public ObjMaterial GetFromFile(ShaderProgram program, string aName, string aFileName, bool FlipY = true)
         {
             if (materials.ContainsKey(aName))
             {
@@ -49,20 +49,21 @@ namespace GameCore.Render.RenderMaterial
             string tempFilePath = Path.Combine(imageDirectory, aFileName);
             if (!File.Exists(tempFilePath))
             {
+                GameCore.TheGameCore.RaiseMessage(string.Format("MaterialManager.GetFromFile() file does not exist: ", tempFilePath));
                 return null;
             }
 
-            Texture tempTexture = new Texture(tempFilePath);
+            Texture tempTexture = new Texture(tempFilePath,FlipY);
             ObjMaterial tempMaterial = new ObjMaterial(program) {DiffuseMap = tempTexture};
             AddMaterial(aName, tempMaterial);
 
             return tempMaterial;
         }
 
-        public ObjMaterial GetFromFile(ShaderProgram program, string aFileName)
+        public ObjMaterial GetFromFile(ShaderProgram program, string aFileName, bool FlipY = true)
         {
             string aName = Path.GetFileNameWithoutExtension(aFileName);
-            return GetFromFile(program, aName, aFileName);
+            return GetFromFile(program, aName, aFileName,FlipY);
         }
 
         public ObjMaterial GetPlainColor(ShaderProgram program, string aName, Color aColor)